+++ apps/app/src/utils.ts
@@ -49,10 +49,19 @@ function formatAddress(address: FormatAddressProps) {
 }
 
 function formatFills(fills: any) {
-  return fills.reduce((prev: string, cur: any) => {
-    const fill = cur.treatment.name;
-    return prev ? `${prev}, ${fill}` : fill;
-  }, '');
+  return fills
+    .reduce((prev: any[], cur: any) => {
+      // Remove duplicates based on prescription id
+      const exists = prev.some((fill) => fill.prescription.id === cur.prescription.id);
+      if (!exists) {
+        prev.push(cur);
+      }
+      return prev;
+    }, [])
+    .reduce((prev: string, cur: any) => {
+      const fill = cur.treatment.name;
+      return prev ? `${prev}, ${fill}` : fill;
+    }, '');
 }
 
 function capitalizeFirst(str: string) {
